Update OpenWeatherMap http request to have API key
This tweaks the URL structure for the OpenWeatherMap HTTP GET request to
have the APPID param with an API key that must be replaced with a real
API key.

The reason for this change is that starting on October 9, 2015,
OpenWeatherMap started to require an API key (a.k.a. APPID) to make
requests. See http://openweathermap.org/faq for more info on this
requirement and http://openweathermap.org/appid on using the APPID.